ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
openai==1.12.0
pymupdf==1.23.21
python-docx==1.1.0
pydantic==2.6.0
httpx==0.26.0
